+++ src/ReceivingList/utils.js
@@ -87,3 +87,16 @@ export const buildTitlesQuery = (queryParams) => {
 
   return connectQuery(filterQuery, sortingQuery);
 };
+
+export const fetchLinesOrders = (mutator, lines, fetchedOrdersMap) => {
+  const unfetched = lines
+    .filter(({ purchaseOrderId }) => !fetchedOrdersMap[purchaseOrderId])
+    .map(({ purchaseOrderId }) => purchaseOrderId)
+    .filter(Boolean);
+
+  const fetchPromise = unfetched.length
+    ? batchFetch(mutator, uniq(unfetched))
+    : Promise.resolve([]);
+
+  return fetchPromise;
+};